More style fixes, add Makefile and dev guide

+++ tests/stats/test_LinearRegression.py
@@ -1,11 +1,9 @@
 """Tests the statistical accuracy of a Linear Regression w/ ProbFlow"""
 
 
-
 import numpy as np
 import tensorflow as tf
 import probflow as pf
-
 
 
 def test_linear_regression():
@@ -18,31 +16,29 @@ def test_linear_regression():
     # Generate data
     N = 1000
     D = 5
-    x = np.random.randn(N, D).astype('float32')
+    x = np.random.randn(N, D).astype("float32")
     w = np.random.randn(D, 1)
     b = np.random.randn()
     std = np.exp(np.random.randn())
-    noise = std*np.random.randn(N, 1)
-    y = x@w + b + noise
-    y = y.astype('float32')
+    noise = std * np.random.randn(N, 1)
+    y = x @ w + b + noise
+    y = y.astype("float32")
 
     # Create and fit model
     model = pf.LinearRegression(D)
     model.fit(x, y, batch_size=100, epochs=1000, lr=1e-2)
 
     # Compute and check confidence intervals on the weights
-    lb, ub = model.posterior_ci('weights')
+    lb, ub = model.posterior_ci("weights")
     assert np.all(lb < w)
     assert np.all(ub > w)
 
     # Compute and check confidence intervals on the bias
-    lb, ub = model.posterior_ci('bias')
+    lb, ub = model.posterior_ci("bias")
     assert lb < b
     assert ub > b
 
     # Compute and check confidence intervals on the std
-    lb, ub = model.posterior_ci('std')
+    lb, ub = model.posterior_ci("std")
     assert lb < std
     assert ub > std
-
-
